# Cranelift: support 14-bit Type index with some bitpacking.