--- a/cranelift/codegen/meta/src/cdsl/types.rs
+++ b/cranelift/codegen/meta/src/cdsl/types.rs
@@ -71,7 +71,7 @@ impl ValueType {
     }
 
     /// Find the unique number associated with this type.
-    pub fn number(&self) -> u8 {
+    pub fn number(&self) -> u16 {
         match *self {
             ValueType::Lane(l) => l.number(),
             ValueType::Reference(r) => r.number(),
@@ -173,7 +173,7 @@ impl LaneType {
     }
 
     /// Find the unique number associated with this lane type.
-    pub fn number(self) -> u8 {
+    pub fn number(self) -> u16 {
         constants::LANE_BASE
             + match self {
                 LaneType::Bool(shared_types::Bool::B1) => 0,
@@ -355,11 +355,11 @@ impl VectorType {
     ///
     /// Vector types are encoded with the lane type in the low 4 bits and
     /// log2(lanes) in the high 4 bits, giving a range of 2-256 lanes.
-    pub fn number(&self) -> u8 {
+    pub fn number(&self) -> u16 {
         let lanes_log_2: u32 = 63 - self.lane_count().leading_zeros();
         let base_num = u32::from(self.base.number());
         let num = (lanes_log_2 << 4) + base_num;
-        num as u8
+        num as u16
     }
 }
 
@@ -411,7 +411,7 @@ impl SpecialType {
     }
 
     /// Find the unique number associated with this special type.
-    pub fn number(self) -> u8 {
+    pub fn number(self) -> u16 {
         match self {
             SpecialType::Flag(shared_types::Flag::IFlags) => 1,
             SpecialType::Flag(shared_types::Flag::FFlags) => 2,
@@ -484,7 +484,7 @@ impl ReferenceType {
     }
 
     /// Find the unique number associated with this reference type.
-    pub fn number(self) -> u8 {
+    pub fn number(self) -> u16 {
         constants::REFERENCE_BASE
             + match self {
                 ReferenceType(shared_types::Reference::R32) => 0,