--- a/cranelift/codegen/src/ir/types.rs
+++ b/cranelift/codegen/src/ir/types.rs
@@ -394,6 +394,18 @@ impl Type {
             self
         }
     }
+
+    /// Gets a bit-level representation of the type. Used only
+    /// internally for efficiently storing types.
+    pub(crate) fn repr(self) -> u16 {
+        self.0
+    }
+
+    /// Converts from a bit-level representation of the type back to a
+    /// `Type`.
+    pub(crate) fn from_repr(bits: u16) -> Type {
+        Type(bits)
+    }
 }
 
 impl Display for Type {