@sparker-arm 2022-06-15
Okay thanks, I will get profiling.