@sparker-arm 2022-06-15
So, _I think__, I've got the overhead down to 0.15%. I've modified `Type::log2_lane_bits` and `Type::log2_lane_count` to return a u32, while modifying `BitSet::contains` to take a u32.

I also removed `ValueData::ty` and directly implemented it on `ValueDataPacked` instead.