@cfallin 2022-07-05
@sparker-arm this is ready for review so that you can rebase the dynamic-vectors work on top of it!